Optimizing speed on dimensionality reducer

+++ test/preProcessing/DimensionalityReducerSpec.scala
@@ -7,12 +7,12 @@ import scala.util.Random
 class DimensionalityReducerSpec extends FlatSpec with Matchers {
   "All outvectors" should "have the same length" in {
     val rnd = new Random(System.currentTimeMillis())
-    val A:DenseMatrix[Double]= DimensionalityReducer.getRandMatrix(20000000,4096);
+    val A:DenseMatrix[Float]= DimensionalityReducer.getRandMatrix(20000000,4096);
 
-    var prev:Array[Double] = Array.empty
+    var prev:Array[Float] = Array.empty
     for(i <- 0 until 100) {
-      val vIn:Array[Double] = Array.fill(4096)(rnd.nextDouble)
-      val vOut:Array[Double] = DimensionalityReducer.getNewVector(vIn,A)
+      val vIn:Array[Float] = Array.fill(4096)(rnd.nextFloat)
+      val vOut:Array[Float] = DimensionalityReducer.getNewVector(vIn,A)
       if(prev != Array.empty) {
         prev.size should be (vOut.size)
       }
@@ -22,10 +22,10 @@ class DimensionalityReducerSpec extends FlatSpec with Matchers {
 
   "Out Vector" should "be of length 220" in {
     val rnd = new Random(System.currentTimeMillis())
-    val A:DenseMatrix[Double]= DimensionalityReducer.getRandMatrix(20000000,4096);
+    val A:DenseMatrix[Float]= DimensionalityReducer.getRandMatrix(20000000,4096);
 
-    val vIn:Array[Double] = Array.fill(4096)(rnd.nextDouble)
-    val vOut:Array[Double] = DimensionalityReducer.getNewVector(vIn,A)
+    val vIn:Array[Float] = Array.fill(4096)(rnd.nextFloat)
+    val vOut:Array[Float] = DimensionalityReducer.getNewVector(vIn,A)
 
     vOut.size should be (220)
     vIn.size should be (4096)
